Unit Test Package Restructure
Restructured unit test package
