+++ PokeDroidSDK/src/main/java/com/geekfed/pokedroid/pokemon/PokemonParser.java
@@ -1,0 +1,13 @@
+package com.geekfed.pokedroid.pokemon;
+
+import org.json.JSONObject;
+
+/**
+ * Created by cody on 5/14/16.
+ */
+public class PokemonParser {
+
+    public PokemonParser(JSONObject pokemonJSON) {
+
+    }
+}
